List the current files in your bucket

The files stored in a bucket are not easily visible from a notebook. List all files in the bucket or files matching a certain pattern.

Usage

aou_ls_bucket(
  pattern = "",
  silent = FALSE,
  recursive = TRUE,
  bucket = Sys.getenv("WORKSPACE_BUCKET"),
  gsutil_args = ""
)

Arguments

pattern

Regular expression, such as "*.csv" or a single file name e.g., "mydata.csv". Default will find all files apart from notebooks (.ipynb files).

silent

Logical; whether to print the names of files found.

bucket

Bucket to retrieve file from. Defaults to `getOption("aou.default.bucket")`, which is `Sys.getenv('WORKSPACE_BUCKET')` unless specified otherwise.

gsutil_args

A string containing other arguments passed to `gsutil ls`. See https://cloud.google.com/storage/docs/gsutil/commands/ls for details.

Value

A vector of file names

Examples

if (FALSE) {
# list all files, including in subdirectories
aou_ls_bucket()
# list all csv files
aou_ls_bucket("*.csv")
# list all csv files in the data directory
aou_ls_bucket("data/*.csv")
}